Implemented bboard::Environment
Implemented an environment class which loosely resembles the
environment from Pommerman.

You can create a game, supply agents to the env and use env.step()

The environment will be used to supply observations to the agents.

You can also env.StartGame(500, true) to render a game for 500 steps
and print out the winner.

Next up:
- Flames need to be ticked and vanished,
- agents need to be able to move into flames

General:
- Enviroment will return observations and rewards to the correct agent