Added Unit Tests for bomb explosions
Didn't test the bombQueue specifically, since I want to test
the specification not the implementation.

Considering Code // Branch coverage, but might be overkill
and less important.

Flames need to be ticked still. Considering a flame queue struct
with a fixed size circular buffer (much like with the bombs).
That would make the flame ticking extremely performant.