Switched to templated FixedQueue for Bombs and Flames
Both flames and bombs use a FixedQueue template, that wraps all
functionality from the previous bombqueue.

FixedQueue uses index shifting on a circular buffer to pop and add
elements.

Flames are correctly added to the queue, next is ticking the flames
in the step and removing the bombs from the actual board (collision
handling will be tricky)