Implemented flame ticking and cleanup
The flames and bombs are both ticked correctly and removed from the
queue by TickFlames and TickBombs. Put them both into step_utilities

Also switched the namespace in step_utilities to bboard::util so it's
easier to see where the methods come from when reading through the step
function.

Next up is removing the flames from the board.