Completed Flame mechanics + unit testing
Flames now don't vanish other flames and work completely correctly.

Changed the Item::, specifically for FLAMES to be 1<<8. This way
the lower 8 bits can be used to identify whose flame it is on the
board.

Added a macro IS_FLAME(), which shifts the value and inspects the 9th bit

All flames are saved in the fixedQueue and there is no need for
additional memory to handle flame collisions now.