First Steps for allowing Bomb Explosion Chains
Currently, as the bomb storage is implemented as a fixed size queue,
only Pop() and Add() are O(1), which is normally not a problem
since Bombs explode in the order they were created on the board.

However that is not the case for chained explosions. This means
bombs can go off at any time.

There are several ways to go about this. I will try to set chained
bombs as "disabled" and see if this can be integrated