Added unit test for adjacent enemies
Currently expanding the strategy module.

Interesting: A simple agent like in pommerman requires dijkstra
(or at least data about reachable nodes). It's possible to create
a very performant list of reachable nodes with an array of
integer bitflags, however the implementation is not straight forward.

The primary concern is creating a SimpleAgent that requires a LOT
less than 100ms to compute its decision. That way we move the bottleneck
to the individual custom agent implementation that might be training
against it.

Next commits will have a couple of new strategy methods like finding
safe movement directions, finding reachable positions etc..